Refactor to use QuestionService only
AnswerService삭제, service 인터페이스 대신 클래스만 사용하도록 수정
